refactor: create ptmalloc2 and jemalloc command categories
