+++ pwndbg/commands/ptmalloc2.py
@@ -11,7 +11,6 @@ from tabulate import tabulate
 
 import pwndbg
 import pwndbg.aglib.heap
-import pwndbg.aglib.heap.jemalloc as jemalloc
 import pwndbg.aglib.memory
 import pwndbg.aglib.proc
 import pwndbg.aglib.typeinfo
@@ -177,7 +176,7 @@ parser.add_argument(
 )
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWhenHeapIsInitialized
 @pwndbg.commands.OnlyWhenUserspace
@@ -228,7 +227,7 @@ parser.add_argument(
 )
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWhenRunning
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWhenHeapIsInitialized
@@ -268,7 +267,7 @@ Default to the current thread's arena.""",
 parser.add_argument("addr", nargs="?", type=int, default=None, help="Address of the arena.")
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWhenHeapIsInitialized
 @pwndbg.commands.OnlyWhenUserspace
@@ -298,7 +297,7 @@ def arena(addr: int | None = None) -> None:
 parser = argparse.ArgumentParser(description="List this process's arenas.")
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWhenHeapIsInitialized
 @pwndbg.commands.OnlyWhenUserspace
@@ -365,7 +364,7 @@ Default to the current thread's tcache.""",
 parser.add_argument("addr", nargs="?", type=int, default=None, help="Address of the tcache.")
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWithTcache
 @pwndbg.commands.OnlyWhenUserspace
@@ -394,7 +393,7 @@ def tcache(addr: int | None = None) -> None:
 parser = argparse.ArgumentParser(description="Print the mp_ struct's contents.")
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWhenHeapIsInitialized
 @pwndbg.commands.OnlyWhenUserspace
@@ -416,7 +415,7 @@ Default to current thread's arena.""",
 parser.add_argument("addr", nargs="?", type=int, default=None, help="Address of the arena.")
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWhenHeapIsInitialized
 @pwndbg.commands.OnlyWhenUserspace
@@ -457,7 +456,7 @@ parser.add_argument(
 )
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWhenHeapIsInitialized
 @pwndbg.commands.OnlyWhenUserspace
@@ -574,7 +573,7 @@ parser.add_argument("addr", nargs="?", type=int, default=None, help="Address of 
 parser.add_argument("tcache_addr", nargs="?", type=int, default=None, help="Address of the tcache.")
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWhenHeapIsInitialized
 @pwndbg.commands.OnlyWhenUserspace
@@ -611,7 +610,7 @@ parser.add_argument(
 )
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWhenHeapIsInitialized
 @pwndbg.commands.OnlyWhenUserspace
@@ -647,7 +646,7 @@ parser.add_argument(
 )
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWhenHeapIsInitialized
 @pwndbg.commands.OnlyWhenUserspace
@@ -683,7 +682,7 @@ parser.add_argument(
 )
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWhenHeapIsInitialized
 @pwndbg.commands.OnlyWhenUserspace
@@ -719,7 +718,7 @@ parser.add_argument(
 )
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWhenHeapIsInitialized
 @pwndbg.commands.OnlyWhenUserspace
@@ -754,7 +753,7 @@ parser.add_argument(
 )
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWithTcache
 @pwndbg.commands.OnlyWhenUserspace
@@ -806,7 +805,7 @@ parser.add_argument(
 )
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWhenHeapIsInitialized
 @pwndbg.commands.OnlyWhenUserspace
@@ -972,7 +971,7 @@ group.add_argument(
 )
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWithResolvedHeapSyms
 @pwndbg.commands.OnlyWhenHeapIsInitialized
 @pwndbg.commands.OnlyWhenUserspace
@@ -1189,7 +1188,7 @@ try_free_parser = argparse.ArgumentParser(
 try_free_parser.add_argument("addr", help="Address passed to free")
 
 
-@pwndbg.commands.ArgparsedCommand(try_free_parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(try_free_parser, category=CommandCategory.PTMALLOC2)
 @pwndbg.commands.OnlyWhenHeapIsInitialized
 @pwndbg.commands.OnlyWhenUserspace
 def try_free(addr: str | int) -> None:
@@ -1561,7 +1560,7 @@ parser.add_argument(
 )
 
 
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
+@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.PTMALLOC2)
 def heap_config(filter_pattern: str) -> None:
     display_config(filter_pattern, "heap", has_file_command=False)
 
@@ -1570,95 +1569,3 @@ def heap_config(filter_pattern: str) -> None:
             "Some config values (e.g. main_arena) will be used only when resolve-heap-via-heuristic is `auto` or `force`"
         )
     )
-
-
-# Jemalloc
-
-parser = argparse.ArgumentParser(
-    description="Returns extent information for pointer address allocated by jemalloc"
-)
-parser.add_argument("addr", type=int, help="Address of the allocated memory location")
-
-
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
-def jemalloc_find_extent(addr) -> None:
-    print(C.banner("Jemalloc find extent"))
-    print("This command was tested only for jemalloc 5.3.0 and does not support lower versions")
-    print()
-
-    addr = int(addr)
-
-    try:
-        rtree = jemalloc.RTree.get_rtree()
-        extent = rtree.lookup_hard(addr)
-        if extent is None:
-            print(message.error("ERROR: Extent not found"))
-            return
-        # print pointer address first, then extent address then extent information
-        print(f"Pointer Address: {hex(addr)}")
-        print(f"Extent Address: {hex(extent.extent_address)}")
-        print()
-
-        jemalloc_extent_info(extent.extent_address, header=False)
-    except pwndbg.dbg_mod.Error as e:
-        print(message.error(f"ERROR: {e}"))
-        return
-
-
-parser = argparse.ArgumentParser(description="Prints extent information for the given address")
-parser.add_argument("addr", type=int, help="Address of the extent metadata")
-parser.add_argument(
-    "-v", "--verbose", action="store_true", help="Print all chunk fields, even unused ones."
-)
-
-
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
-def jemalloc_extent_info(addr, verbose=False, header=True) -> bool:
-    if header:
-        print(C.banner("Jemalloc extent info"))
-        print("This command was tested only for jemalloc 5.3.0 and does not support lower versions")
-        print()
-
-    try:
-        extent = jemalloc.Extent(int(addr))
-
-        print(f"Allocated Address: {hex(extent.allocated_address)}")
-        print(f"Extent Address: {hex(extent.extent_address)}")
-
-        print(f"Size: {hex(extent.size)}")
-        print(f"Small class: {extent.has_slab}")
-
-        print(f"State: {extent.state_name}")
-
-        if verbose:
-            for bit, val in extent.bitfields.items():
-                print(bit, val)
-    except pwndbg.dbg_mod.Error as e:
-        print(message.error(f"ERROR: {e}"))
-        return False
-    return True
-
-
-parser = argparse.ArgumentParser(description="Prints all extents information")
-
-
-@pwndbg.commands.ArgparsedCommand(parser, category=CommandCategory.HEAP)
-def jemalloc_heap() -> None:
-    print(C.banner("Jemalloc heap"))
-    print("This command was tested only for jemalloc 5.3.0 and does not support lower versions")
-    print()
-
-    try:
-        rtree = jemalloc.RTree.get_rtree()
-        extents = rtree.extents
-        if len(extents) == 0:
-            print(message.warn("No extents found"))
-            return
-        for extent in extents:
-            # TODO: refactor so not create copies
-            if not jemalloc_extent_info(extent.extent_address, header=False):
-                return
-            print()
-    except pwndbg.dbg_mod.Error as e:
-        print(message.error(f"ERROR: {e}"))
-        return
